# openapi.yaml
openapi: 3.0.3
info:
  title: API ИИ-гида музея Фаберже
  version: 0.1.0-mvp
  description: |
    Backend-API мобильного web-приложения (PWA) «ИИ-гид музея Фаберже».

    Посетитель сканирует **QR-код** на входе экспозиции, попадает в приложение
    без регистрации и установки, перемещается по **интерактивной карте**
    (зал → витрина → экспонат), а главное — может **сфотографировать экспонат**,
    получить о нём рассказ от ИИ-гида и **прослушать** его озвучку.

    Этот документ описывает контракт backend-сервиса (FastAPI на Yandex Cloud),
    к которому обращается фронтенд. Спецификация написана по дорожной карте MVP
    (25 мая – 25 июня) и служит единым контрактом между Frontend, Backend и ML.

    ## Ключевой пользовательский сценарий (бесшовное распознавание)

    ```
    Клик «Распознать экспонат»
      → камера (WebRTC)
      → снимок
      → POST /recognition (фото → YOLO → label_slug)
      → POST /guide/story (label_slug → raw_history из БД → YandexGPT → рассказ)
      → показ рассказа + 3–4 всплывающих вопроса
      → POST /speech («Прослушать», SpeechKit → аудио)
      → блок «Другие экспонаты зала» (GET /exhibits/{id}/related)
    ```

    ## Доменная модель

    Иерархия каталога повторяет физическую структуру музея:

    | Уровень   | Сущность   | Таблица БД   | Ключевые поля                                   |
    |-----------|------------|--------------|-------------------------------------------------|
...

...
    (через публичный API не отдаётся).

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
paths:
...
  /map:
    get:
      tags: [Карта и навигация]
      summary: Интерактивная карта музея
      description: >-
        Полное дерево экспозиции для отрисовки интерактивной карты:
        залы с вложенными витринами и счётчиками экспонатов.
        Один запрос на загрузку карты на первом экране.
        Описание зала приходит и полным (`description`), и превью
        (`description_preview` + `description_has_more`) — см. `GET /halls/{hall_id}`.
      operationId: getMap
      security: []
      parameters:
        - $ref: '#/components/parameters/IsTemporary'
        - $ref: '#/components/parameters/IncludeService'
      responses:
        '200':
          description: Дерево залов и витрин.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapResponse'
  /halls:
    get:
      tags: [Карта и навигация]
      summary: Список залов
      description: >-
        Постраничный список залов музея для карты и навигации. Служебные записи
        (`is_service`) в выдачу не попадают — источник правды один, скрывать их
        на клиенте не нужно. Описание зала приходит и полным (`description`), и
        превью (`description_preview` + `description_has_more`) — см.
        `GET /halls/{hall_id}`.
      operationId: listHalls
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IsTemporary'
        - $ref: '#/components/parameters/IncludeService'
      responses:
        '200':
          description: Список залов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallListResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /halls/{hall_id}:
    get:
      tags: [Карта и навигация]
      summary: Получить зал
      description: >-
        Карточка зала с описанием и кнопкой «Спросить AI-гида» на фронте.
        Описание зала используется на уровне зала интерактивной карты.


        На первом экране показывается `description_preview`, а полный
        `description` раскрывается кнопкой «Подробнее о зале» — её показывать,
        только когда `description_has_more = true`. Так на одной странице видно
        и что есть текст о зале, и что есть витрины: раньше описание занимало
        весь экран и витрины уезжали за него (п. I-3 баг-репорта 31.08.2026).
        Второго запроса раскрытие не требует — полный текст уже в ответе.
      operationId: getHall
      security: []
      parameters:
        - $ref: '#/components/parameters/HallId'
      responses:
        '200':
          description: Данные зала.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /halls/{hall_id}/showcases:
    get:
      tags: [Карта и навигация]
      summary: Витрины зала
      description: Список витрин внутри указанного зала.
      operationId: listHallShowcases
      security: []
      parameters:
        - $ref: '#/components/parameters/HallId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список витрин зала.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /halls/{hall_id}/exhibits:
    get:
      tags: [Карта и навигация]
      summary: Экспонаты зала
      description: >-
        Все экспонаты зала (по всем его витринам). Эта же выборка лежит в
        основе блока «Другие экспонаты зала» — соседние записи по `hall_id`.
      operationId: listHallExhibits
      security: []
      parameters:
        - $ref: '#/components/parameters/HallId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список экспонатов зала.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /showcases:
    get:
      tags: [Карта и навигация]
      summary: Список / фильтрация витрин
      description: >-
        Плоский постраничный список витрин по всем залам, с необязательным
        фильтром по `hall_id`. Универсальная выборка для админки и навигации
        (формат ответа — как у `/halls` и `/exhibits`).
      operationId: listShowcases
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: hall_id
          in: query
          description: Фильтр по идентификатору зала.
          required: false
          schema:
            type: integer
            minimum: 1
          example: 5
      responses:
        '200':
          description: Список витрин.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseListResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /showcases/{showcase_id}:
    get:
      tags: [Карта и навигация]
      summary: Получить витрину
      operationId: getShowcase
      security: []
      parameters:
        - $ref: '#/components/parameters/ShowcaseId'
      responses:
        '200':
          description: Данные витрины.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /showcases/{showcase_id}/exhibits:
    get:
      tags: [Карта и навигация]
      summary: Экспонаты витрины
      operationId: listShowcaseExhibits
      security: []
      parameters:
        - $ref: '#/components/parameters/ShowcaseId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список экспонатов витрины.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ─────────────────────────── Экспонаты ─────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Статический Bearer-токен администратора (получается через
        `POST /admin/login`). Слать в заголовке `Authorization: Bearer <token>`
        ко всем эндпоинтам `/admin/**` (кроме `/admin/login`).
  parameters:
    Limit:
      name: limit
      in: query
      description: Максимум элементов на странице.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Смещение от начала выборки (постраничность).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    IsTemporary:
      name: is_temporary
      in: query
      required: false
      description: >-
        Фильтр по типу экспозиции: `true` — временные (сменные) выставки,
        `false` — основная (постоянная) экспозиция. Без параметра возвращаются
        и те и другие. Для экспонатов признак берётся от зала (через витрину).
      schema:
        type: boolean
      example: true
    IncludeService:
      name: include_service
      in: query
      required: false
      description: >-
        Включить служебные записи каталога (`is_service`) — технические залы и
        вспомогательные пространства. По умолчанию `false` — посетителю такие
        записи не показываются. `true` нужен админке, чтобы служебный зал
        оставался управляемым.
      schema:
        type: boolean
        default: false
      example: false
    HallId:
      name: hall_id
      in: path
      required: true
      description: Идентификатор зала.
      schema:
        type: integer
        minimum: 1
      example: 4
    ShowcaseId:
      name: showcase_id
      in: path
      required: true
      description: Идентификатор витрины.
      schema:
        type: integer
        minimum: 1
      example: 12
...
  responses:
    NotFound:
      description: Ресурс не найден.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Экспонат не найден.
...
    UnprocessableEntity:
      description: Ошибка валидации запроса (формат FastAPI).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
...
  schemas:
...
    Hall:
      type: object
      # description_preview/description_has_more считает бэкенд, поэтому в ответе
      # они есть ВСЕГДА — как и в схеме, которую генерирует FastAPI.
      required: [id, description_preview, description_has_more]
      properties:
        id:
          type: integer
          example: 4
        hall_number:
          type: integer
          nullable: true
          description: >-
            Номер зала на плане экспозиции (уникальный). `null` — у зала номера
            нет («Вне постоянной экспозиции»): подпись строится только из
            названия, без «зал № …». Такие залы выводятся последними.
          example: 4
        name:
          type: string
          nullable: true
          example: Синяя гостиная
        description:
          type: string
          nullable: true
          description: >-
            Описание зала для уровня зала на карте и для контекста ИИ-гида.
            Наполняется из плана экспозиции и путеводителя.
          example: >-
            Парадная гостиная Шуваловского дворца, где представлены девять
            императорских пасхальных яиц Фаберже.
        level:
          type: integer
          nullable: true
          description: Этаж / уровень в здании музея.
          example: 2
        cover_image_url:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.cloud/halls/blue/cover.jpg
        is_temporary:
          type: boolean
          default: false
          description: >-
            Зал временной (сменной) выставки. `true` — временная выставка,
            `false` — основная (постоянная) экспозиция. Основа для тега/фильтра
            «Временная выставка» в каталоге и переключателя в админке.
          example: false
        is_service:
          type: boolean
          default: false
          description: >-
            Служебная запись каталога — технические залы и вспомогательные
            пространства. Такие залы не попадают ни в `GET /halls`, ни на карту,
            ни в ответы гида — увидеть их можно только с
            `?include_service=true` (нужно админке).
            Зал №1 «Парадная лестница» был служебным с 29.07 по 31.08.2026;
            с 31.08.2026 он — обычный первый зал основной экспозиции, и в
            публичной выдаче стало 12 пронумерованных залов вместо 11.
          example: false
        sort_order:
          type: integer
          default: 0
          description: Порядок вывода в каталоге/админке (drag-n-drop).
          example: 0
        showcase_count:
          type: integer
          nullable: true
          example: 6
        exhibit_count:
          type: integer
          nullable: true
          example: 24
        description_preview:
          type: string
          nullable: true
          readOnly: true
          description: >-
            Видимая часть текста о зале: начало `description`, обрезанное ПО
            ГРАНИЦЕ ПРЕДЛОЖЕНИЯ (не по символу и не по слову) до
            `HALL_DESCRIPTION_PREVIEW_CHARS` знаков — по умолчанию 350 при
            описаниях в каталоге на 1,8–4,1 тыс. знаков. Полный текст никуда не
            делся и лежит рядом в `description`: за раскрытием второй запрос не
            нужен. `null` только если `null` само описание. Если превью
            выключено (`HALL_DESCRIPTION_PREVIEW_CHARS=0`), равно `description`.
          example: >-
            Рыцарский зал был отделан по проекту Бернара де Симона 1844 года и
            частично изменён в 1859 году Рудольфом Бернгардом.
        description_has_more:
          type: boolean
          default: false
          readOnly: true
          description: >-
            Есть ли что раскрывать: `true` — `description` длиннее превью,
            показывать «Подробнее о зале»; `false` — описание уместилось целиком
            (`description_preview == description`), кнопки быть не должно.
            Признак считает бэкенд — фронту не нужно сравнивать строки.
          example: true
    HallBrief:
      type: object
      description: Краткая ссылка на зал (для вложений в карточку экспоната).
      required: [id]
      properties:
        id:
          type: integer
          example: 4
        hall_number:
          type: integer
          nullable: true
          description: '`null` — зал без номера («Вне постоянной экспозиции»).'
          example: 4
        name:
          type: string
          nullable: true
          example: Синяя гостиная
    HallDetail:
      allOf:
        - $ref: '#/components/schemas/Hall'
        - type: object
          properties:
            showcases:
              type: array
              description: Витрины зала (могут включаться по необходимости).
              items:
                $ref: '#/components/schemas/Showcase'
    MapHall:
      allOf:
        - $ref: '#/components/schemas/Hall'
        - type: object
          properties:
            showcases:
              type: array
              items:
                $ref: '#/components/schemas/Showcase'
    MapResponse:
      type: object
      required: [halls]
      properties:
        halls:
          type: array
          items:
            $ref: '#/components/schemas/MapHall'
    HallListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Hall'
        total:
          type: integer
          example: 12
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Витрины ──
    Showcase:
      type: object
      required: [id, hall_id]
      properties:
        id:
          type: integer
          example: 12
        hall_id:
          type: integer
          example: 4
        showcase_number:
          type: integer
          nullable: true
          description: >-
            Номер витрины внутри зала (уникален в пределах зала). `null` —
            группа «не в витринах»: в путеводителе она отмечена пустым
            квадратом, это экспонаты зала вне витрин. Такая группа в зале одна
            (попытка завести вторую → `409`) и выводится последней.
          example: 5
        name:
          type: string
          nullable: true
          example: Императорские пасхальные яйца
        exhibit_count:
          type: integer
          nullable: true
          example: 4
...
    ShowcaseDetail:
      allOf:
        - $ref: '#/components/schemas/Showcase'
        - type: object
          properties:
            hall:
              allOf:
                - $ref: '#/components/schemas/HallBrief'
              nullable: true
              description: 'Зал витрины; `null` — витрина не привязана к залу.'
            exhibits:
              type: array
              items:
                $ref: '#/components/schemas/ExhibitSummary'
    ShowcaseListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Showcase'
        total:
          type: integer
          example: 6
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Экспонаты ──
...
    ExhibitLocation:
      type: object
      description: >-
        Готовое расположение предмета (п. I-2 баг-репорта 31.08.2026). Объект
        присутствует ВСЕГДА; если экспонат не привязан к витрине, все поля `null`
        — ключи при этом не исчезают. `text` — строка для показа
        («Зал 4 «Синяя гостиная», витрина 5»), `text_in` — та же фраза в обороте
        «Найти его можно в зале 4 «Синяя гостиная», витрина 5». Формулировка
        совпадает с той, что печатает ИИ-гид (общий модуль `app/services/location.py`),
        поэтому собирать её из `hall_*`/`showcase_*` на клиенте не нужно. Числа и
        идентификаторы оставлены для ссылок «перейти в зал / в витрину».
      required:
        [hall_id, hall_number, hall_name, showcase_id, showcase_number, showcase_name, text, text_in]
      properties:
        hall_id:
          type: integer
          nullable: true
          example: 4
        hall_number:
          type: integer
          nullable: true
          description: '`null` — зал без номера («Вне постоянной экспозиции»): в подписи только название.'
          example: 4
        hall_name:
          type: string
          nullable: true
          example: Синяя гостиная
        showcase_id:
          type: integer
          nullable: true
          example: 12
        showcase_number:
          type: integer
          nullable: true
          description: '`null` — группа «не в витринах» (в путеводителе — пустой квадрат); в тексте это «вне витрин».'
          example: 5
        showcase_name:
          type: string
          nullable: true
          description: Название витрины, если музей его ведёт.
          example: Пасхальные подарки
        text:
          type: string
          nullable: true
          description: Готовая строка расположения для показа отдельной плашкой над названием.
          example: Зал 4 «Синяя гостиная», витрина 5
        text_in:
          type: string
          nullable: true
          description: Та же фраза в предложном падеже — для оборота «Найти его можно …».
          example: в зале 4 «Синяя гостиная», витрина 5
    ExhibitMaker:
      type: object
      description: >-
        Фирма и мастер (п. I-2 баг-репорта 31.08.2026). `text` — поле `master_name`
        дословно, оно ВСЕГДА авторитетно и его достаточно для показа;
        `firm`/`master` — его разобранные части (эвристика по формулировкам
        путеводителя), обе `null`, если разобрать не удалось. Разбор производный:
        источник правды в БД один — `master_name`, отдельных колонок под фирму и
        мастера нет и они не планируются, пока музей не утвердит форму записи фирм.
        Части возвращаются дословными подстроками `text`, вместе со словом-маркером
        («Фирма», «Фабрика», «мастер») и кавычками источника.
      required: [text, firm, master]
      properties:
        text:
          type: string
          nullable: true
          example: Фирма К. Фаберже, мастер М. Перхин
        firm:
          type: string
          nullable: true
          example: Фирма К. Фаберже
        master:
          type: string
          nullable: true
          example: мастер М. Перхин
    ExhibitSummary:
      type: object
      description: >-
        Краткое представление экспоната для списков, поиска и карусели. Порядок
        свойств — тот же, что в `Exhibit` (см. его описание): плашка «зал/витрина»
        над названием рисуется одинаково в каталоге и в карточке. Набор ключей
        постоянен — пустое приходит как `null`, а не отсутствием ключа.
      required:
        [id, exhibit_number, label_slug, name, thumbnail_url, location, hall_id, showcase_id,
         showcase_number, year_created, maker, master_name, is_temporary]
      properties:
        id:
          type: integer
          example: 101
        exhibit_number:
          type: string
          nullable: true
          description: >-
            Номер экспоната по путеводителю музея (B3). Показывается перед
            названием в каталоге. Строка — в путеводителе бывают номера вида «12а».
          example: '12'
        label_slug:
          type: string
          nullable: true
          example: faberge_egg_coronation
        name:
          type: string
          example: Яйцо «Коронационное»
        thumbnail_url:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.cloud/exhibits/coronation/thumb.jpg
        location:
          allOf:
            - $ref: '#/components/schemas/ExhibitLocation'
          description: >-
            Готовое расположение — то же самое, что в карточке. Добавлено
            31.08.2026: раньше в списке был только `hall_id`, без названия зала,
            и плашку «зал/витрина» собрать было не из чего.
        hall_id:
          type: integer
          nullable: true
          description: 'Дубль `location.hall_id`; оставлен для совместимости.'
          example: 4
        showcase_id:
          type: integer
          nullable: true
          description: 'Дубль `location.showcase_id`; оставлен для совместимости.'
          example: 12
        showcase_number:
          type: integer
          nullable: true
          description: >-
            Номер витрины экспоната (`null` — вне витрин). Отдаётся прямо в
            списке, чтобы группировку «витрина → её экспонаты» можно было
            собрать из одного ответа `GET /halls/{hall_id}/exhibits`.
            Дубль `location.showcase_number`.
          example: 5
        year_created:
          type: string
          nullable: true
          description: >-
            Датировка как в путеводителе («1899–1903», «1890-е», «конец XIX
            века») — подпись под названием берите отсюда. До 17.08.2026 поле
            было числом (нижняя граница) и жило в паре с дублем `dating`;
            теперь поле датировки одно. Техник (`techniques`), материалов и
            места создания в списке нет — они только в карточке `GET /exhibits/{id}`.
          example: '1899–1903'
        maker:
          allOf:
            - $ref: '#/components/schemas/ExhibitMaker'
          description: Фирма и мастер разобранными частями; полная строка — в `maker.text`.
        master_name:
          type: string
          nullable: true
          description: 'Дубль `maker.text`; оставлен для совместимости.'
          example: Фирма К. Фаберже, мастер М. Перхин
        is_temporary:
          type: boolean
          nullable: true
          description: >-
            Признак временной выставки, унаследованный от зала экспоната
            (через витрину → зал). Позволяет показать бейдж «Временная выставка»
            в карточке каталога.
          example: false
...
    ExhibitListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitSummary'
        total:
          type: integer
          example: 24
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Поиск ──
...
    Error:
      type: object
      description: Стандартная ошибка приложения (FastAPI `HTTPException`).
      required: [detail]
      properties:
        detail:
          type: string
          example: Экспонат не найден.
    ValidationErrorItem:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          example: [body, message]
        msg:
          type: string
          example: field required
        type:
          type: string
          example: value_error.missing
    HTTPValidationError:
      type: object
      description: Ошибка валидации запроса в формате FastAPI.
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorItem'